jarvis-ai/requirements.txt
# JARVIS AI Operating Layer - Dependencies
# Version: 1.0.0

# ============ Core AI ============
ollama>=0.1.0
# transformers>=4.35.0  # Optional for advanced features

# ============ Voice - STT ============
openai-whisper>=20230314
# faster-whisper>=0.9.0  # Alternative, faster

# ============ Voice - TTS ============
piper-tts>=1.0.0

# ============ Voice - Wake Word ============
# pvporcupine>=2.0.0  # Requires API key
openwakeword>=0.5.0  # Free alternative

# ============ Audio ============
pyaudio>=0.2.12
sounddevice>=0.4.6
soundfile>=0.12.1
webrtcvad>=2.0.10
numpy>=1.24.0

# ============ Memory & Vectors ============
chromadb>=0.4.0
sentence-transformers>=2.0.0

# ============ Database ============
sqlalchemy>=2.0.0
# sqlcipher3>=0.5.0  # For encrypted storage

# ============ System Control ============
pyautogui>=0.9.54
pynput>=1.7.6
psutil>=5.9.0
pywin32>=306
pyperclip>=1.8.2
mss>=9.0.0

# ============ Screen & Vision ============
pillow>=10.0.0
opencv-python>=4.8.0
# paddleocr>=2.7.0  # For OCR (large, install separately)
# pytesseract>=0.3.10  # Alternative OCR

# ============ Browser Automation ============
playwright>=1.40.0

# ============ Security ============
cryptography>=41.0.0

# ============ Web Server (UI) ============
fastapi>=0.104.0
uvicorn>=0.24.0

# ============ Utilities ============
python-dotenv>=1.0.0
apscheduler>=3.10.0
rich>=13.0.0
pydantic>=2.0.0

# ============ Development ============
pytest>=7.4.0
black>=23.0.0
gitpython>=3.1.0
